Control.Monad.Trail

Description

Trail type which makes an ordinary Monad out of FMonad

Documentation

newtype Trail (mm :: (Type -> Type) -> Type -> Type) a

For any FMonad mm, Trail mm is a Monad.

Example

Trail mm can become variantions of Monad for different FMonad mm.

mm = Compose m

For any Monad m, Trail (Compose m) is isomorphic to m.

Trail (Compose m) a
  ~ Compose m ((,) a) ()
  ~ m (a, ())
  ~ m a

mm = Free

Trail Free is isomorphic to the list monad [].

Trail Free a
  ~ Free ((,) a) ()
  ~ [a]

mm = FreeT' m

For any Monad m, Trail (FreeT' m) is isomorphic to ListT m, where ListT is so-called "ListT done right."

Trail (FreeT' m) a
  ~ FreeT ((,) a) m ()
  ~ ListT m a

See more for examples/ListTVia.hs

Constructors

Trail

Fields

runTrail :: mm ((,) a) ()