Data.Functor.Precompose

Documentation

type (:.:) = Compose :: (Type -> Type) -> (Type -> Type) -> Type -> Type

Single-kinded type alias of Compose

newtype Precompose (f :: j -> k) (g :: k -> Type) (a :: j)

Flipped-order Compose.

When f is a Monad, Precompose f is a FMonad in the similar way Compose is.

The only difference is Precompose f composes f to the right (_pre_compose) compared to Compose f which composes to the left (_post_compose).

Constructors

Precompose

Fields

getPrecompose :: g (f a)